refactor: decompose 4 monolith files into focused submodules (issue #2486 sprint 1)
Split 4 files exceeding 800-1020 LOC into smaller focused modules while
preserving backward-compatible re-exports. Each coordinator file now delegates
to purpose-built submodules under 500 LOC.

- filters.py (836 LOC) → filters/ package: filter_design, smoothing_filters,
  advanced_filters, shortcuts
- output_manager.py (899 → 495 LOC) → _format_handlers, _path_utils,
  _report_generators, _async_io, _simulation_store
- ball_flight_physics.py (896 → 64 LOC) → ball_properties, ball_launch_conditions,
  ball_simulator, ball_enhanced_simulator, ball_trajectory_analysis
- urdf_io.py (1020 → 102 LOC) → urdf_constants, urdf_generator, urdf_parser
